+++ src/services/query-string.ts
@@ -1,10 +1,18 @@
-// Parse a query string into an object
-export function parseQueryString(string: string) {
-  if (string == '') {
-    return {}
-  }
-  const segments = string.split('&').map((s) => s.split('='))
-  const queryString: Record<string, unknown> = {}
-  segments.forEach((s) => (queryString[s[0]] = s[1]))
-  return queryString
+export function formatQueryString(obj: Record<string, unknown>): string {
+  const params = Object.entries(obj).reduce<URLSearchParams>(
+    (memo, [key, value]) => {
+      if (value === undefined || value === null) {
+        return memo
+      }
+
+      if (Array.isArray(value)) {
+        value.forEach((val) => memo.append(key, val))
+      } else {
+        memo.append(key, String(value))
+      }
+      return memo
+    },
+    new URLSearchParams(),
+  )
+  return params.toString()
 }