+++ src/submission-service.ts
@@ -5,7 +5,9 @@ import {
   getPendingQueueSubmissions,
   getFormSubmission,
   updatePendingQueueSubmission,
+  cancelEditingPendingQueueSubmission,
   deletePendingQueueSubmission,
+  editPendingQueueSubmission,
   registerPendingQueueAttachmentProgressListener,
   registerPendingQueueProgressListener,
   executePendingQueueProgressListeners,
@@ -71,6 +73,14 @@ async function processPendingQueue({
     if (isOffline()) {
       console.log(
         'Application is offline, leaving submission in the pending queue:',
+        pendingQueueSubmission,
+      )
+      continue
+    }
+
+    if (pendingQueueSubmission.isEditing) {
+      console.log(
+        'submission is being edited, leaving submission in the pending queue:',
         pendingQueueSubmission,
       )
       continue
@@ -509,6 +519,8 @@ export {
   goBackOrCloseWindow,
   getPendingQueueSubmissions,
   deletePendingQueueSubmission,
+  editPendingQueueSubmission,
+  cancelEditingPendingQueueSubmission,
   registerPendingQueueListener,
   processPendingQueue,
   BaseFormSubmission,