+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `submissionService.editPendingQueueSubmission`
+- `submissionService.cancelEditingPendingQueueSubmission`
+
 ## [14.0.1] - 2024-08-07
 
 ### Dependencies